% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_classification.R
\name{byPhenotype}
\alias{byPhenotype}
\title{Outputs list of matrices, each containing samples from a single disease phenotype
and the reference condition.}
\usage{
byPhenotype(input_matrix, phenotypes, reference_condition)
}
\arguments{
\item{input_matrix}{A matrix of discriminating scores for pathway pairs,
where each row represents a sample and each columns is a pair of pathway.}

\item{phenotypes}{A vector of strings containing the phenotype associated with each sample.
Must be the same length as the number of rows in the feature_matrix input.}

\item{reference_condition}{String. The phenotype identifier for the reference condition in *phenotypes*.
For example, 'GFP'.}
}
\description{
Splits a given matrix of discriminating scores by phenotype.
The output can be used to generate a binary classification model.
}
