% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_networkCharacterization.R
\name{characterizeNetworks}
\alias{characterizeNetworks}
\title{Characterize reactome pathway networks.}
\usage{
characterizeNetworks(networks_list, output_dir = NULL)
}
\arguments{
\item{networks_list}{A named list of igraph networks, where the names are the
subtypes represented in each network.}

\item{output_dir}{The directory to which the xlsx files will be saved.}
}
\description{
Saves xlsx files containing the common names for Reactome pathways
# representing nodes in a given pathway network.
}
