% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_classification.R
\name{crosstalkNetwork}
\alias{crosstalkNetwork}
\title{Generate a network containing the top-performing pathway pairs.}
\usage{
crosstalkNetwork(
  crosstalk_matrix,
  groups,
  alpha = 1,
  lambda = "best",
  output_network = TRUE,
  model_evaluation = FALSE
)
}
\arguments{
\item{crosstalk_matrix}{A matrix of discriminating scores in which each column
is a pair of enriched pathways and each row is a sample.
The output of pathwayCrosstalk().}

\item{groups}{A dataframe containing the mappings between sample identifiers
('sample_id', a factor with the reference condition as the first level) and
associated treatment conditions ('group'). The sample identifiers must be in
the same order as the columns of the count_matrix.}

\item{alpha}{The alpha value applied to the glmnet classification model.
Use alpha = 1 for lasso regression and alpha = 0 for ridge regression.}

\item{lambda}{The lambda value applied to the glmnet classification model.}

\item{output_network}{Logical. If TRUE, the function outputs the network object.}

\item{model_evaluation}{Logical. If TRUE, the function outputs the model evaluation metrics.}
}
\value{
Depending on the values of the `output_network` and `model_evaluation`
arguments, either a list containing the network object or a list containing
the model evaluation metrics.
}
\description{
Generates a glmnet binomial classification model for a given matrix of
pathway-pair discriminating scores and outputs a network object containing the
top-performing pathway pairs in the model.
}
