% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_pathwayEnrichment.R
\name{.fisherPathwayEnrichment}
\alias{.fisherPathwayEnrichment}
\title{Do pathway enrichment with Fisher's exact test}
\usage{
.fisherPathwayEnrichment(deg, gene_alpha)
}
\arguments{
\item{deg}{diffExpression --> entrez_to_hgnc}

\item{gene_alpha}{numeric vector indicating significance level for genes.}
}
\value{
data frame of results for the Fisher's exact tests.
    `p` Unadjusted p-value
    `adj_p` FDR adjusted p-value
    `estimate` odds ratio point estimate
    `pathway` Paste of Reactome pathway identifier and description.
}
\description{
Do pathway enrichment with Fisher's exact test
}
