% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayEnrichment.R
\name{fisherPathwayEnrichment}
\alias{fisherPathwayEnrichment}
\title{Pathway enrichment using Fisher's exact test}
\usage{
fisherPathwayEnrichment(deg, alpha, pathways)
}
\arguments{
\item{deg}{Output of diffExpression function.}

\item{alpha}{numeric vector indicating significance level for the differential expression analysis. Genes in the differential
expression analysis with an adjusted p-value less than alpha are considered to be differentially expressed for the purpose
of doing Fisher's exact test for every pathway.}

\item{pathways}{Named list where names are pathway identifiers and elements are sets of Entrez gene identifiers (character)}
}
\value{
data frame of results for the Fisher's exact tests.
    `p` Unadjusted p-value.
    `adj_p` FDR adjusted p-value.
    `estimate` odds ratio point estimate.
    `pathway` pathway identifier.
}
\description{
Pathway enrichment using Fisher's exact test
}
