% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_crosstalkInhibition.R
\name{networkEfficiency}
\alias{networkEfficiency}
\title{Conduct simulated network crosstalk inhibition and prune network.}
\usage{
networkEfficiency(network)
}
\arguments{
\item{network}{An igraph object in which the nodes represent enriched pathways and
the edges represent significant pathway crosstalks. The output of crosstalkNetwork().}
}
\value{
Returns a named list.
     \itemize{
     \item full_network - An igraph object representing the full
      network of top-performing pathway pairs.
     \item full_network_results - A dataframe describing the full network.
     \item pruned_network - An igraph object representing the pruned
      network of top-performing pathway pairs.
     \item pruned_network_results - A dataframe describing the pruned network.
     }
}
\description{
Given a network, calculate the network efficiency. Iterate over
network edges and conduct simulated network crosstalk inhibition, pruning the network
of edges that do not contribute to network efficiency.
}
