% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_networkWrapper.R
\name{networkWrapper}
\alias{networkWrapper}
\title{networkWrapper}
\usage{
networkWrapper(
  expression_matrix,
  groups,
  DEPs,
  pathways,
  pathway_alpha,
  lambda,
  ...
)
}
\arguments{
\item{expression_matrix}{An matrix of gene expression data where the
row names are the gene probe identifiers and column names are the sample identifiers.
For RNAseq data, a matrix of un-normalized integer counts.
For microarray data, a matrix of intensity values for gene probes with
pre-processing completed, including normalization, removing control sequences,
log2 transformation, and collapse to genes.}

\item{groups}{A dataframe containing the sample identifiers ('sample_id', a factor with
the reference condition as the first level) and associated treatment conditions ('group').
The sample identifiers must be in the same order as the columns of the count_matrix.}

\item{DEPs}{A dataframe of pathways with associated p-values in a column 'p'.
The output of pathwayEnrichment().}

\item{pathways}{A list in which each element is a named list of genes
corresponding to a particular pathway.}

\item{pathway_alpha}{Significance level for enriched pathways. Pathways with a Fisher's exact
test p-value less than `pathway_alpha` will be labeled as enriched for the purposes
of downstream analyses.}

\item{...}{Additional arguments to crosstalkNetwork() function.}
}
\value{
Returns a named list.
     \itemize{
     \item full_network - An igraph object representing the full
      network of top-performing pathway pairs.
     \item full_network_results - A dataframe describing the full network.
     \item pruned_network - An igraph object representing the pruned
      network of top-performing pathway pairs.
     \item pruned_network_results - A dataframe describing the pruned network.
     }
}
\description{
Wrapper function for the pathwayTalk workflow.
}
