% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_pathwayCrosstalk.R
\name{pathwayCrosstalk}
\alias{pathwayCrosstalk}
\title{Calculate matrix of discriminating scores for pathway crosstalk}
\usage{
pathwayCrosstalk(enriched_pathways, treatment_map, processes = 4)
}
\arguments{
\item{enriched_pathways}{Output of fisherPathwayEnrichment}
}
\description{
Calculate matrix of discriminating scores for pathway crosstalk
}
