% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_classification.R
\name{selectTopFeatures}
\alias{selectTopFeatures}
\title{Allows for selection of features with greatest variation.}
\usage{
selectTopFeatures(input_matrix, n_top_features = 20000)
}
\arguments{
\item{input_matrix}{A matrix of discriminating scores for pathway pairs,
where each row represents a sample and each columns is a pair of pathways.}

\item{n_top_features}{Integer. The number of top features returned.}
}
\description{
Selects the features of a given input matrix with the greatest variance.
The n_top_features argument allows for the specification of the number of features returned.
Returns a matrix containing only the top features of the original matrix.
}
