% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_classification.R
\name{subtypeNetwork}
\alias{subtypeNetwork}
\title{Outputs a graph object of the top-performing pathway pairs for a given subtype.}
\usage{
subtypeNetwork(
  feature_matrix,
  sample_phenotype,
  alpha = 1,
  lambda = "best",
  output_graph = TRUE,
  model_evaluation = FALSE
)
}
\arguments{
\item{feature_matrix}{A matrix of discriminating scores for pathway pairs,
where each row represents a sample and each columns is a pair of pathways.
#' The matrix should include samples from the reference condition and one disease phenotype.}

\item{sample_phenotype}{A vector of strings containing the phenotype associated with each sample.
Must be the same length as the number of rows in the feature_matrix input.}

\item{alpha}{The alpha value applied to the glmnet classification model.
Use alpha = 1 for lasso regression and alpha = 0 for ridge regression.}

\item{lambda}{The lambda value applied to the glmnet classification model.}

\item{output_graph}{Logical. If TRUE, the function outputs the graph object.}

\item{model_evaluation}{Logical. If TRUE, the function outputs the model evaluation metrics.}
}
\description{
Generates a glmnet binomial classification model for a given matrix of
pathway-pair discriminating scores.
Outputs a graph object containing the top-performing pathway pairs in the model.
}
