% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_diffExpression.R
\name{diffExpression}
\alias{diffExpression}
\title{Conduct differential gene expression analysis for a given gene expression dataset.}
\usage{
diffExpression(
  data,
  collapse_method = "MaxMean",
  gene_ids,
  design_mat,
  contrast_mat
)
}
\arguments{
\item{data}{An expression matrix of intensity values for gene probes,
where the row names are the probe identifiers.}

\item{collapse_method}{Specifies the method by which the probes will be collapsed
to gene expression. The default value is 'MaxMean'. See the WGCNA collapseRows()
documentation for more options.}

\item{gene_ids}{A vector of strings containing gene IDs corresponding to each row of the
expression matrix.}

\item{design_mat}{A matrix specifying the experimental design,
generated with model.matrix().
#' @param contrast_mat A matrix specifying the preferred contrasts, generated
with makeContrasts().}
}
\description{
Collapses probe-level expression to genes using collapseRows()
from WGCNA, and calculates log2 fold-changes and associated p-values for a
given matrix of gene expression data.
}
