% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_pathwayEnrichment.R
\name{fisherPathwayEnrichment}
\alias{fisherPathwayEnrichment}
\title{fisherPathwayEnrichment}
\usage{
fisherPathwayEnrichment(DEGs, gene_alpha, pathways)
}
\arguments{
\item{DEGs}{A dataframe of differential gene expression results in with
associated p-values in a column 'pvalue'. Extracted from the output of diffExpression().}

\item{gene_alpha}{Significance level for differentially expressed genes. Genes with a
p-value less than `gene_alpha` will be labeled as differentially expressed for the purpose
doing a Fisher's exact test for pathway enrichment.}

\item{pathways}{A named list in which each element is a pathway containing a
character vector of the corresponding gene IDs.}
}
\value{
A data frame of results for the Fisher's exact tests.
 \itemize{
 \item p - Unadjusted p-value.
 \item adj_p - FDR adjusted p-value.
 \item estimate - Odds ratio point estimate.
 \item pathway - Pathway ID.
 }
}
\description{
Conduct Fisher's test for pathway enrichment analysis for a set
of differentially expressed genes.
}
