% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayCrosstalk.R
\name{pathwayCrosstalk}
\alias{pathwayCrosstalk}
\title{Calculate matrix of discriminating scores for pathway crosstalk}
\usage{
pathwayCrosstalk(enriched_pathways, genes, pathways, processes = 1)
}
\arguments{
\item{enriched_pathways}{Character vector of enriched pathways}

\item{genes}{Output of WGCNA::collapseRows}

\item{pathways}{List where names are pathway names and elements are sets of Entrez
gene identifiers.}

\item{processes}{Number of processes (integer) that should be used}
}
\value{
Matrix of dimensions $m x n$ where $m$ is the number of pathway pairs and $n$ is the number
    of samples in the data. Each element of this matrix is the discriminating score for a particular sample,
    which is defined as:
    $\frac{mean(g_x) - mean(g_y)}{sd(g_x) - sd(g_y)}$
}
\description{
Calculate matrix of discriminating scores for pathway crosstalk
}
